commit 29/08/2024

+++ src/app/pages/absenceTools/absence-search/absence-search.component.ts
@@ -3,6 +3,8 @@ import {NgForOf} from "@angular/common";
 import {RouterLink, RouterLinkActive} from "@angular/router";
 import {AbsenceService} from "../service/absence.service";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
+import {ToastrService} from "ngx-toastr";
+import {iconApp, manager} from "../../../models/models";
 
 @Component({
   selector: 'app-absence-search',
@@ -24,7 +26,7 @@ export class AbsenceSearchComponent implements OnInit {
   private show: boolean = false;
   keyword:  string = "";
 
-  constructor(protected service: AbsenceService){}
+  constructor(protected service: AbsenceService , private toastr: ToastrService){}
 
   ngOnInit(): void {
   }
@@ -36,22 +38,29 @@ export class AbsenceSearchComponent implements OnInit {
     this.service.searchAbsence(keyword).subscribe(data =>{
         this.absences = data
         console.log(data)
+        if(this.absences.length == 0) this.toastr.warning(iconApp+" Aucune Absence ne corresponds à ce mot clé !! " , manager , {enableHtml:true});
     } , error => {
       console.log(error);
+      this.toastr.warning(iconApp+ `ce genre de caractère n'est pas permit : ${keyword} ` , manager , {enableHtml:true});
     })
   }
 
   deleteAbsence(id : number) {
+    let conf  =  confirm("Cette absence sera supprimée !!")
+
+    if(!conf) return;
+
     this.service.deleteAbsence(id).subscribe(data =>{
       console.log(data);
-      this.show = true ;
+      this.toastr.success(iconApp + " Cette absence a été supprimé avec succès!! " , manager , {enableHtml:true});
 
+
+    } , error => {
+      console.log(error);
+      this.toastr.error(iconApp + " Une erreur est survenue lors de la suppression !!!"  ,manager , {enableHtml : true})
     })
   }
-  reloadNotification() {
-    this.show = false ;
-    window.location.reload();
-  }
+
