+++ src/app/pages/congetools/conge-list/conge-list.component.ts
@@ -1,7 +1,9 @@
 import {Component, OnInit} from '@angular/core';
 import {CongeService} from "../service/conge.service";
 import {NgForOf, NgIf} from "@angular/common";
-import {RouterLink, RouterLinkActive} from "@angular/router";
+import {Router, RouterLink, RouterLinkActive} from "@angular/router";
+import {ToastrService} from "ngx-toastr";
+import {iconApp, manager} from "../../../models/models";
 
 @Component({
   selector: 'app-conge-list',
@@ -21,7 +23,7 @@ export class CongeListComponent implements OnInit{
 
   protected header : string[] = ["No" , "début de congé" , "fin de congé" , "type de congé" , "status" , "demande de" , "actions"]
 
-  constructor(protected service : CongeService ){ }
+  constructor(protected service : CongeService   , private toastr : ToastrService ) { }
 
 
   ngOnInit(): void {
@@ -31,8 +33,13 @@ export class CongeListComponent implements OnInit{
 
   deleteConge(id : number) {
 
+
+    let conf = confirm("Are you sure you want to delete this conge?");
+    if(!conf) return;
     this.service.deleteConge(id).subscribe(data =>{
       console.log(data);
+      this.toastr.success(iconApp+" congé supprimé avec succès...." , manager , {enableHtml:true});
+
     })
   }
 
@@ -42,6 +49,7 @@ export class CongeListComponent implements OnInit{
       this.listConge = data;
     } , error => {
       console.log(error);
+      this.toastr.warning(iconApp+ " Une erreur est survenue au !!");
 
     })
   }
