+++ src/app/pages/contenuTools/contenu-search/contenu-search.component.ts
@@ -1,9 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {ContenuService} from "../service/contenu.service";
-import {ContenuModel} from "../../../models/models";
+import {ContenuModel, iconApp, manager} from "../../../models/models";
 import {FormsModule} from "@angular/forms";
 import {NgForOf} from "@angular/common";
 import {RouterLink, RouterLinkActive} from "@angular/router";
+import {ToastrService} from "ngx-toastr";
 
 @Component({
   selector: 'app-contenu-search',
@@ -23,7 +24,7 @@ export class ContenuSearchComponent implements OnInit{
   protected contenuPicked: ContenuModel[] = [] ;
 
   protected  header = ["No" , "Titre" , "Thème" , "Nature" , "Langue" , "Date de création" , "status" , "Actions"]
-  constructor(protected service : ContenuService) {
+  constructor(protected service : ContenuService , private toastr : ToastrService) {
   }
 
   ngOnInit(): void {
@@ -32,11 +33,14 @@ export class ContenuSearchComponent implements OnInit{
   searchContenu(keyword: string){
       this.service.searchContenu(keyword).subscribe (data => {
           this.contenuPicked = data;
+          if(data.length == 0 ) this.toastr.warning(iconApp + " Aucun contenu ne correspond a ce mot clé !! " , manager , {enableHtml:true});
       } , error => {
         console.log(error);
       })
   }
   deleteContenu(id:number) {
+    let conf = confirm("Ce contenu sera supprimer !!")
+    if(conf) return ;
     this.service.deleteContenu(id).subscribe (data => {
       console.log(data);
       window.location.reload();
